+++ src/solver/optimisation/optim_post_process_list.cpp
@@ -26,7 +26,7 @@ OptPostProcessList::OptPostProcessList(PROBLEME_HEBDO* problemeHebdo,
                                                                            resultWriter));
     post_process_list.push_back(
       std::make_unique<DumpSimulationTablePostProcessCmd>(problemeHebdo_,
-                                                          IO::Outputs::Stage::remixHydro,
+                                                          IO::Outputs::Stage::peakShaving,
                                                           simulationTables));
     post_process_list.push_back(
       std::make_unique<InterpolateWaterValuePostProcessCmd>(problemeHebdo_, areas, calendar));